requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2025.1.31
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.1.0
frozenlist==1.6.0
greenlet==3.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.3
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
redis==5.2.1
six==1.17.0
SQLAlchemy==2.0.40
SQLAlchemy-Utils==0.41.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
